Get dynamic status type when generating plot "OBO Principle by status" in analysis page (#118)
* move analysis notebook to py code first draft

* Define dynamic status type to generate dataframe

The code expected to have the four status types in the dashboard
results. This is not always the case, so the code was updated to check
for the available status types before reshaping the data.

* Script and jinja2 template to generate analysis.html

The script processes the dashboard results and prepares it for analysis.
It generates plots and tables for the analysis page. The script replaces
the dashboard_analysis.ipynb Jupyter notebook with some improvements.

* replaces rule to generate analysis.html with python script

The analysis.html page is now generated by a python script instead of
the jupyter notebook.

---------

Co-authored-by: Anita Caron <anitacaron@users.noreply.github.com>