Skip oboscore recalculation if ontology unchanged (#117)
* skip obo score recalculation if ontology unchanged

* fix condition for when skip is true and ontology has changed or not

Change the condition to verify the `dashboard.yml` has changed instead
of check only if dashboard.yml exists for each ontology,. In case of
previous failure, it runs again. Before, the config `skip_existing`,
according to the log message, was expecting the `dashboard.yml` existence,
but it was outside the first condition where it checks for the file. Now,
if there isn't a `dashboard.yml` for an ontology and `skip_existing`
is true, it won't skip.

---------

Co-authored-by: Anita Caron <anitacaron@users.noreply.github.com>
Co-authored-by: Anita Caron <anitacaron@Anitas-MacBook-Pro.local>
Co-authored-by: Anita Caron <antic@ebi.ac.uk>