generalise the report generation for ROBOT report and FP3 and FP7 reports
When fixing the incompleteness of ROBOT report generation, I didn't
realise it was also used for the FP3 and FP7 reports, so I made it more
general now.
Change default value for `limitlines` to 50 when generating FP reports
instead changing in two places in the makefile.
Only assert error_count_level and error_count_rule when it's the ROBOT report
Add condition in the template to show types of error and error breakdown
only when the it's ROBOT report
Generalise the row values in the table
Only change row colour accordingly to error level when it's the ROBOT report